% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acdcourse-package.R
\docType{package}
\name{acdcourse-package}
\alias{acdcourse}
\alias{acdcourse-package}
\title{acdcourse: Analyzing Cohort Datasets in R}
\description{
Four tutorial chapters covering the basics on how to analyze cohort 
    datasets using R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lwjohnst86/acdcourse}
  \item Report bugs at \url{https://github.com/lwjohnst86/acdcourse/issues}
}

}
\author{
\strong{Maintainer}: Luke Johnston \email{lwjohnst@gmail.com} (0000-0003-4169-2616)

}
\keyword{internal}
